LayoutTests/imported/w3c/web-platform-tests/webrtc-encoded-transform/encoded-frame-clone.https.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <meta name="timeout" content="long">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="video1" autoplay></video>
        <video id="video2" autoplay></video>
        <script src ="routines.js"></script>
        <script>
async function getInboundRTPStats(receiver)
{
    const report = await receiver.getStats();
    var stats;
    report.forEach((statItem) => {
        if (statItem.type === "inbound-rtp") {
            stats = statItem;
        }
    });
    return stats;
}

async function getDecodedFrameCount(receiver)
{
    const stats = await getInboundRTPStats(receiver);
    if (!stats)
        return 0;
    return stats.kind === "video" ? stats.framesDecoded : stats.jitterBufferEmittedCount;
}

async function checkDecodedFrameCountIsIncreasing(test, receiver, testName)
{
    const frameCount = await getDecodedFrameCount(receiver);
    let counter = 0;
    do {
        await waitFor(test, 200); 
        const newFrameCount = await getDecodedFrameCount(receiver);
        if (newFrameCount !== frameCount)
            break;
    } while (++counter < 20);

    assert_less_than(counter, 20, "checkDecodedFrameCountIsIncreasing " + (testName ? testName : ""));
}

async function checkDecodedFrameCountIsFreezing(test, receiver, testName)
{
    let frameCount = await getDecodedFrameCount(receiver);
    let counter = 0;
    do {
        await waitFor(test, 200); 
        const newFrameCount = await getDecodedFrameCount(receiver);
        if (newFrameCount === frameCount)
            break;
        frameCount = newFrameCount;
    } while (++counter < 20);

    assert_less_than(counter, 20, "checkDecodedFrameCountIsFreezing " + (testName ? testName : ""));
}

async function checkUsingClone(test, port, receiver)
{
    await checkDecodedFrameCountIsIncreasing(test, receiver, "initial");

    port.postMessage("startCloningFrame");
    await checkDecodedFrameCountIsFreezing(test, receiver);

    port.postMessage("stopCloningFrame");
    await checkDecodedFrameCountIsIncreasing(test, receiver, "final");
}

promise_test(async (test) => {
    const {sender, receiver, senderPc, receiverPc} = await createConnectionWithTransform(test, 'encoded-frame-worker.js', {video: true});
    return checkUsingClone(test, sender.transform.port, receiver);
}, "Clone frames in sender RTCEncodedVideoFrame");

promise_test(async (test) => {
    const {sender, receiver, senderPc, receiverPc} = await createConnectionWithTransform(test, 'encoded-frame-worker.js', {video: true});
    return checkUsingClone(test, receiver.transform.port, receiver);
}, "Clone frames in receiver RTCEncodedVideoFrame");

promise_test(async (test) => {
    const {sender, receiver, senderPc, receiverPc} = await createConnectionWithTransform(test, 'encoded-frame-worker.js', {audio: true});
    return checkUsingClone(test, sender.transform.port, receiver);
}, "Clone frames in sender RTCEncodedAudioFrame");

promise_test(async (test) => {
    const {sender, receiver, senderPc, receiverPc} = await createConnectionWithTransform(test, 'encoded-frame-worker.js', {audio: true});
    return checkUsingClone(test, receiver.transform.port, receiver);
}, "Clone frames in receiver RTCEncodedAudioFrame");
       </script>
    </body>
</html>
